스테이터스UI/ 장비랑 플레이어 스테이터스연동 + 스탯 증가 버튼이 경험치바랑 연동되는 버그있음
왜 생긴건지도 모르겠음

+++ Assets/Scripts/UI/Others/HPBar.cs
@@ -5,8 +5,8 @@ public class HPBar : MonoBehaviour
 {
     public GameObject hpBar;
 
-    public int curHp;
-    public int maxHp;
+    public float curHp;
+    public float maxHp;
     public Image hpImg;
 
     private void Awake()
